refactor: consolidated modals (create/update) into one, and renamed add-ranking modal.
